common, msg: lower messenger's debug gathering level 5 -> 0.
The change is a result of profiling a `vstart.sh`-deployed cluster.
When the `debug_ms` is kept default, `msgr-worker` threads spend
plenty of cycles just on crafting debugs.

```
  Children      Self  Command        Shared Object         Symbol
-    8,49%     0,08%  msgr-worker-0  ceph-osd              [.] operator<<
   - 8,42% operator<<
      - 7,27% operator<<
         - 6,18% getnameinfo
            + 4,51% inet_ntop
            + 1,41% __snprintf
           0,62% std::__ostream_insert<char, std::char_traits<char> >
      - 0,96% std::ostream::_M_insert<unsigned long>
```

Additional issue is that the `Log::submit_entry()` has a critical
section, and thus serializes execution.

The patch radically deals with the problems by lowering the default
gathering level of the `ms` subsystem.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>