rgw: switch bufferlist::copies related to RGWGetDataCB::handle_data().
These users are relatively straightforward to judge whether the last_p
was in effect for them or not. The reason behind that is the fact that
all of them are called from

RGWGetDataCB

get_obj_data::flush():

RGWGetObj_Decompress::handle_data() – where the `out_bl` is
a subject to bufferlist::splice() between iteration. Effectively
this zeroizes the last_p each time.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>