osd: let users of get_omap_iterator() select between upper and lower_bound()
This squeezes additional calling into OMAP iterator on the hot
RGW bucket listing path while preserving the pure-extend
characteristic of the interface modification; that is, there
is no 2nd collection locking in BlueStore while the patch
should be easier (in terms of risk) to backport.

Inspired by cxyxd's PR 59384 and Adam Kupczyk's PR 60056.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>